Fixes #12
At least sort of?  It would be really hard to enforce in some ways...
so, I did this - basically, /ber reset on - turns on reset requirement.
The next bid for EVERYONE will be accompanied with a warning that it
will cost all points.  If they win that auction (or auctions) they LOSE
ALL POINTS.  Even if they win only one of two (multiple bids is where it
would get complicated to do any other way).  Basically... if you're
going to split bids and it's a reset condition... careful!

It also marks people's named with a * when they are in a reset condition
in status and DKP listing.