Affects... #25, #17
Basically refactored DKP granting a little bit, mostly to deal with
automatically granting it.  Also... attempted to totally redo
attendance/loot taking.  Unfortunately, it didn't get saved for some
reason.  I might not have fully reloaded the addon, as the TOC file does
not get reloaded on reloadui.  Might need to restart the game for that,
damn it.  I wanted to see it!  Oh well, tomorrow...