Fixes #11
Implemented minimum bids.  Untested so far.  Think it should work.
Basically, it warns players when they bid less than the set minimum bid
and less than their total (bidding their total, if under minimum, is
fine).  Then when subtracting DKP on a win, it subtracts off the minimum
or their total DKP, whichever is lower.  If the bid is above minimum,
it just uses that still of course.