Fixes #46
Well... removed the old automatic/crappy import functionality.  It was
really a oneshot, and it should never ever be used again.  Added a /ber
import command now.  It imports from GRSS_Data for the same DKP name
set.  Seems to work pretty well.  Created a button for it on the UI as
well so it's easy to do.